requirements.txt
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==5.1.1
defusedxml==0.7.1
executing==2.0.1
fire==0.5.0
gitdb==4.0.11
GitPython==3.1.41
idna==3.6
ipython==8.21.0
jaraco.classes==3.3.1
jedi==0.19.1
jira==3.6.0
keyring==24.3.0
matplotlib-inline==0.1.6
more-itertools==10.2.0
oauthlib==3.2.2
packaging==23.2
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-dotenv==1.0.1
python-gitlab==4.4.0
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
termcolor==2.4.0
traitlets==5.14.1
typing_extensions==4.9.0
urllib3==2.2.0
wcwidth==0.2.13
